+++ R/testthat.R
@@ -1,4 +1,4 @@
-#' Expectation: shinytest object passed snapshot tests
+#' Expectation: `testApp()` passes snapshot tests
 #'
 #' This returns an testthat expectation object.
 #'
@@ -10,35 +10,29 @@
 #' \dontrun{
 #' expect_pass(testApp("path/to/app/"))
 #' }
-#'
 #' @export
 expect_pass <- function(object, info = NULL) {
   if (!inherits(object, "shinytest.results")) {
-    testthat::expect(
-      FALSE,
-      "expect_pass requires results from shinytest::testApp()."
-    )
+    stop("expect_pass() requires results from shinytest::testApp()", call. = FALSE)
   }
 
   pass_idx <- vapply(object$results, `[[`, "pass", FUN.VALUE = FALSE)
   fail_names <- vapply(object$results[!pass_idx], `[[`, "name", FUN.VALUE = "")
 
   all_pass <- all(pass_idx)
-  diff_txt <- ""
-
   if (!all_pass) {
     diff_txt <- textTestDiff(object$appDir, fail_names, object$images)
-  }
-
-  testthat::expect(
-    all_pass,
-    sprintf(
-      "Not all shinytest scripts passed for %s: %s\n\nDiff output:\n%s",
+    message <- sprintf(
+      "Not all shinytest scripts passed for %s: %s\n\nDiff output:\n%s\n%s",
       object$appDir,
       paste(fail_names, collapse = ", "),
-      diff_txt
-    ),
-    info = info
-  )
+      diff_txt,
+      paste0("If this is expected, use `snapshotUpdate('", object$appDir, "')` to update")
+    )
+  } else {
+    message <- ""
+  }
+
+  testthat::expect(all_pass, message, info = info)
   invisible(object)
 }