+++ tests/testthat/recorded_tests/009-upload/tests/shinytest/mytest.R
@@ -1,4 +1,4 @@
-app <- ShinyDriver$new("../")
+app <- ShinyDriver$new("../../")
 app$snapshotInit("mytest")
 
 app$uploadFile(file1 = "mtcars.csv")