+++ tests/testthat/recorded_tests/app-waitForValue/tests/shinytest/mytest.R
@@ -1,4 +1,4 @@
-app <- ShinyDriver$new("../")
+app <- ShinyDriver$new("../../")
 app$snapshotInit("mytest")
 
 # Check the box to display the slider and plot